+++ com-mmxlabs-common-util/com.mmxlabs.common.util/src/main/java/com/mmxlabs/common/curves/InterpolatingDiscountCurve.java
@@ -14,16 +14,16 @@ import java.util.TreeMap;
  * 
  */
 public class InterpolatingDiscountCurve implements ICurve {
-	final TreeMap<Integer, Double> values = new TreeMap<Integer, Double>();
+	final TreeMap<Integer, Integer> values = new TreeMap<Integer, Integer>();
 
-	public void setValueAtPoint(final int time, final double discountValue) {
+	public void setValueAtPoint(final int time, final int discountValue) {
 		values.put(time, discountValue);
 	}
 
 	@Override
-	public double getValueAtPoint(final double point) {
-		final Entry<Integer, Double> above = values.ceilingEntry((int) point);
-		final Entry<Integer, Double> below = values.floorEntry((int) point);
+	public int getValueAtPoint(final int point) {
+		final Entry<Integer, Integer> above = values.ceilingEntry(point);
+		final Entry<Integer, Integer> below = values.floorEntry(point);
 
 		if (above == null) {
 			if (below == null) {
@@ -46,11 +46,11 @@ public class InterpolatingDiscountCurve implements ICurve {
 	 * @param point
 	 * @return
 	 */
-	private double interpolate(final Entry<Integer, Double> below, final Entry<Integer, Double> above, final double point) {
+	private int interpolate(final Entry<Integer, Integer> below, final Entry<Integer, Integer> above, final int point) {
 		if (below == above) {
 			return below.getValue();
 		}
-		final double d = above.getKey() - below.getKey();
+		final int d = above.getKey() - below.getKey();
 		return below.getValue() + (((above.getValue() - below.getValue()) * (above.getKey() - point)) / d);
 	}
